+++ frontend/src/pages/tracks/Tracks.js
@@ -2,15 +2,18 @@ import React, { useEffect, useState } from "react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
 import axios from "axios";
 
-import { Button, Table } from "react-bootstrap";
+import Button from "react-bootstrap/Button";
+import Container from "react-bootstrap/Container";
+import Table from "react-bootstrap/Table";
 
 import Track from "./Track.js";
 import ConfirmModal from "../../components/ConfirmModal.js";
 import Spinner from "../../components/Spinner/Spinner.js";
 
 import mainClasses from "../pages.module.scss";
+import isAuthorized from "../../util/isAuthorized.js";
 
-function Tracks({ setAlert }) {
+function Tracks({ setAlert, canEdit = false }) {
   //////////////
   // useHooks //
   //////////////
@@ -36,11 +39,13 @@ function Tracks({ setAlert }) {
   return (
     <div className={mainClasses["fade-in"]}>
       <Spinner loading={loading} />
-      <div className="container my-5">
-        <h1 className="m-0">Tracks</h1>
-      </div>
-      {search && <Track setAlert={setAlert} reFetch={fetchData} />}
-      <div className="container">
+      <Container>
+        <h1>Tracks</h1>
+      </Container>
+      {search && (
+        <Track setAlert={setAlert} reFetch={fetchData} canEdit={canEdit} />
+      )}
+      <Container>
         <ConfirmModal
           id={id}
           showModal={showModal}
@@ -52,59 +57,76 @@ function Tracks({ setAlert }) {
           confirmLabel="Yes, Delete track"
         />
         {response && response.length && (
-          <Table bordered hover striped variant="dark">
-            <thead>
-              <tr>
-                <th>Title</th>
-                <th>Release</th>
-                <th>Year</th>
-                <th></th>
-              </tr>
-            </thead>
-            <tbody>
-              {Array.isArray(response) &&
-                response.map((track) => {
-                  const { id, title, release, releaseId, year } = track;
-                  return (
-                    <tr key={id}>
-                      <td className="position-relative">
-                        <Link to={`?id=${id}`} className="stretched-link">
-                          {title}
-                        </Link>
-                      </td>
-                      <td>
-                        <Link to={`/releases?id=${releaseId}`}>{release}</Link>
-                      </td>
-                      <td>{year}</td>
-                      <td style={{ width: "0" }}>
-                        <Button
-                          size="sm"
-                          variant="outline-danger"
-                          className="text-nowrap text-light"
-                          onClick={() => {
-                            setId(id);
-                            handleShowModal(true);
-                          }}
-                        >
-                          Delete track
-                        </Button>
-                      </td>
-                    </tr>
-                  );
-                })}
-            </tbody>
-          </Table>
+          <>
+            <TableHeader />
+            <Table bordered hover striped variant="dark">
+              <thead>
+                <tr>
+                  <th>Title</th>
+                  <th style={{ width: "0", whiteSpace: "nowrap" }}>
+                    Track No.
+                  </th>
+                  <th>Release</th>
+                  <th>Year</th>
+                  {canEdit && <th></th>}
+                </tr>
+              </thead>
+              <tbody>
+                {Array.isArray(response) &&
+                  response.map((track) => {
+                    const { trackId, id, title, release, releaseId, year } =
+                      track;
+                    return (
+                      <tr key={id}>
+                        <td className="position-relative">
+                          <Link to={`?id=${id}`} className="stretched-link">
+                            {title}
+                          </Link>
+                        </td>
+                        <td>{trackId}</td>
+                        <td className="position-relative">
+                          <Link
+                            to={`/releases?id=${releaseId}`}
+                            className="stretched-link"
+                          >
+                            {release}
+                          </Link>
+                        </td>
+                        <td>{year}</td>
+                        {canEdit && (
+                          <td style={{ width: "0" }}>
+                            <Button
+                              disabled={!canEdit}
+                              size="sm"
+                              variant="danger"
+                              className="text-nowrap text-light"
+                              onClick={() => {
+                                setId(id);
+                                handleShowModal(true);
+                              }}
+                            >
+                              Delete track
+                            </Button>
+                          </td>
+                        )}
+                      </tr>
+                    );
+                  })}
+              </tbody>
+            </Table>
+          </>
         )}
-        {response && (
+        {response && canEdit && (
           <Button
+            disabled={!canEdit}
             onClick={handleAdd}
-            variant="outline-primary"
-            className="text-light"
+            variant="primary"
+            className="text-light mb-5"
           >
             Create new track
           </Button>
         )}
-      </div>
+      </Container>
     </div>
   );
 
@@ -117,6 +139,10 @@ function Tracks({ setAlert }) {
   }
 
   async function handleDelete(id) {
+    const isAuthorizedUser = isAuthorized(canEdit, setAlert);
+    if (!isAuthorizedUser) {
+      return;
+    }
     try {
       setLoading(true);
       const response = await axios.delete(`${pathname}?id=${id}`);
@@ -156,6 +182,17 @@ function Tracks({ setAlert }) {
         setLoading(false);
       }
     })();
+  }
+
+  /////////////////
+  // TableHeader //
+  /////////////////
+
+  function TableHeader() {
+    if (!search) {
+      return null;
+    }
+    return <h3>All tracks</h3>;
   }
 }
